Changes BLE implementation to use the interface
