+++ targets/ESP32/_nanoCLR/nanoFramework.Device.Bluetooth/sys_dev_ble_native.cpp
@@ -41,8 +41,6 @@ static const CLR_RT_MethodHandler method_lookup[] =
     NULL,
     NULL,
     NULL,
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattReadRequest::NativeReadRespondWithValue___VOID__U2__SZARRAY_U1,
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattReadRequest::NativeReadRespondWithProtocolError___VOID__U2__U1,
     NULL,
     NULL,
     NULL,
@@ -78,9 +76,6 @@ static const CLR_RT_MethodHandler method_lookup[] =
     NULL,
     NULL,
     NULL,
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattWriteRequest::NativeWriteGetData___SZARRAY_U1__U2,
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattWriteRequest::NativeWriteRespond___VOID__U2,
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattWriteRequest::NativeWriteRespondWithProtocolError___VOID__U2__U1,
     NULL,
     NULL,
     NULL,
@@ -105,7 +100,6 @@ static const CLR_RT_MethodHandler method_lookup[] =
     NULL,
     NULL,
     NULL,
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattLocalCharacteristic::NativeNotifyClient___I4__U2__U2__SZARRAY_U1,
     NULL,
     NULL,
     NULL,
@@ -357,9 +351,6 @@ static const CLR_RT_MethodHandler method_lookup[] =
     NULL,
     NULL,
     NULL,
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattServiceProvider::NativeInitService___BOOLEAN,
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattServiceProvider::NativeStartAdvertising___BOOLEAN,
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattServiceProvider::NativeStopAdvertising___VOID,
     NULL,
     NULL,
     NULL,
@@ -391,6 +382,33 @@ static const CLR_RT_MethodHandler method_lookup[] =
     NULL,
     NULL,
     NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip::NativeInitService___BOOLEAN,
+    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip::NativeStartAdvertising___BOOLEAN,
+    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip::NativeStopAdvertising___VOID,
+    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip::NativeNotifyClient___I4__U2__U2__SZARRAY_U1,
+    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip::NativeReadRespondWithValue___VOID__U2__SZARRAY_U1,
+    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip::NativeReadRespondWithProtocolError___VOID__U2__U1,
+    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip::NativeWriteGetData___SZARRAY_U1__U2,
+    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip::NativeWriteRespond___VOID__U2,
+    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip::NativeWriteRespondWithProtocolError___VOID__U2__U1,
     NULL,
     NULL,
     NULL,
@@ -432,7 +450,7 @@ static const CLR_RT_MethodHandler method_lookup[] =
 const CLR_RT_NativeAssemblyData g_CLR_AssemblyNative_nanoFramework_Device_Bluetooth =
 {
     "nanoFramework.Device.Bluetooth",
-    0x935F7800,
+    0x47F78837,
     method_lookup,
     { 100, 0, 0, 3 }
 };
