+++ targets/ESP32/_nanoCLR/nanoFramework.Device.Bluetooth/sys_dev_ble_native.h
@@ -10,9 +10,7 @@
 #include <nanoCLR_Interop.h>
 #include <nanoCLR_Runtime.h>
 #include <nanoPackStruct.h>
-
-#include <string.h>
-
+#include <corlib_native.h>
 #include "sys_ble.h"
 
 typedef enum __nfpack BluetoothError
@@ -113,6 +111,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_BluetoothDevice
     static const int FIELD___id = 1;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_BluetoothEvent
@@ -123,6 +122,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_BluetoothEvent
     static const int FIELD__descriptorId = 6;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_Buffer
@@ -131,6 +131,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_Buffer
     static const int FIELD___length = 2;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattLocalDescriptorParameters
@@ -140,6 +141,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___staticValue = 3;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattPresentationFormat
@@ -151,33 +153,37 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___description = 5;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattReadRequest
 {
     static const int FIELD___readValue = 1;
     static const int FIELD___eventID = 2;
-
-    NANOCLR_NATIVE_DECLARE(NativeReadRespondWithValue___VOID__U2__SZARRAY_U1);
-    NANOCLR_NATIVE_DECLARE(NativeReadRespondWithProtocolError___VOID__U2__U1);
+    static const int FIELD___nativeDevice = 3;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattReadRequestedEventArgs
 {
     static const int FIELD___eventID = 1;
     static const int FIELD___session = 2;
+    static const int FIELD___nativeDevice = 3;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattWriteRequestedEventArgs
 {
     static const int FIELD___eventID = 1;
     static const int FIELD___session = 2;
+    static const int FIELD___nativeDevice = 3;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattLocalDescriptor
@@ -192,6 +198,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD__WriteRequested = 8;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattClientNotificationResult
@@ -202,6 +209,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___bytesSent = 4;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattSubscribedClient
@@ -209,6 +217,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___session = 1;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattSession
@@ -216,6 +225,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___deviceId = 1;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattWriteRequest
@@ -224,12 +234,10 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___offset = 2;
     static const int FIELD___value = 3;
     static const int FIELD___eventID = 4;
-
-    NANOCLR_NATIVE_DECLARE(NativeWriteGetData___SZARRAY_U1__U2);
-    NANOCLR_NATIVE_DECLARE(NativeWriteRespond___VOID__U2);
-    NANOCLR_NATIVE_DECLARE(NativeWriteRespondWithProtocolError___VOID__U2__U1);
+    static const int FIELD___nativeDevice = 5;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattLocalCharacteristic
@@ -244,25 +252,28 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___properties = 6;
     static const int FIELD___descriptors = 7;
     static const int FIELD___subscribedClients = 8;
-    static const int FIELD___userDescription = 9;
-    static const int FIELD___userDescriptionDescriptor = 10;
-    static const int FIELD___presentationFormats = 11;
-    static const int FIELD___presentationFormatsDescriptors = 12;
-    static const int FIELD___staticValue = 13;
-    static const int FIELD__ReadRequested = 14;
-    static const int FIELD__WriteRequested = 15;
-    static const int FIELD__SubscribedClientsChanged = 16;
-
-    NANOCLR_NATIVE_DECLARE(NativeNotifyClient___I4__U2__U2__SZARRAY_U1);
+    static const int FIELD___nativeDevice = 9;
+    static const int FIELD___userDescription = 10;
+    static const int FIELD___userDescriptionDescriptor = 11;
+    static const int FIELD___presentationFormats = 12;
+    static const int FIELD___presentationFormatsDescriptors = 13;
+    static const int FIELD___staticValue = 14;
+    static const int FIELD__ReadRequested = 15;
+    static const int FIELD__WriteRequested = 16;
+    static const int FIELD__SubscribedClientsChanged = 17;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_BluetoothEventListener
 {
     static const int FIELD_STATIC___characteristicMap = 1;
 
+    static const int FIELD___nativeDevice = 1;
+
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_DataReader
@@ -271,10 +282,10 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_DataReader
     static const int FIELD___currentReadPosition = 2;
 
     //--//
+
 };
 
-struct
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattLocalCharacteristicParameters
+struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattLocalCharacteristicParameters
 {
     static const int FIELD___writeProtectionLevel = 1;
     static const int FIELD___readProtectionLevel = 2;
@@ -284,6 +295,7 @@ struct
     static const int FIELD___presentationFormats = 6;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattLocalCharacteristicResult
@@ -292,6 +304,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___error = 2;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattLocalDescriptorResult
@@ -300,18 +313,20 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___descriptor = 2;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattLocalService
 {
     static const int FIELD___serviceUuid = 1;
     static const int FIELD___characteristics = 2;
+    static const int FIELD___nativeDevice = 3;
 
     //--//
+
 };
 
-struct
-    Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattServiceProviderAdvertisingParameters
+struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattServiceProviderAdvertisingParameters
 {
     static const int FIELD___deviceName = 1;
     static const int FIELD___isDiscoverable = 2;
@@ -319,6 +334,7 @@ struct
     static const int FIELD___serviceData = 4;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattServiceProvider
@@ -326,17 +342,15 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD_STATIC___bluetoothEventManager = 2;
 
     static const int FIELD___services = 1;
-    static const int FIELD___status = 2;
-    static const int FIELD___deviceName = 3;
-    static const int FIELD___isDiscoverable = 4;
-    static const int FIELD___isConnectable = 5;
-    static const int FIELD___serviceData = 6;
-
-    NANOCLR_NATIVE_DECLARE(NativeInitService___BOOLEAN);
-    NANOCLR_NATIVE_DECLARE(NativeStartAdvertising___BOOLEAN);
-    NANOCLR_NATIVE_DECLARE(NativeStopAdvertising___VOID);
+    static const int FIELD___nativeDevice = 2;
+    static const int FIELD___status = 3;
+    static const int FIELD___deviceName = 4;
+    static const int FIELD___isDiscoverable = 5;
+    static const int FIELD___isConnectable = 6;
+    static const int FIELD___serviceData = 7;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttributeProfile_GattServiceProviderResult
@@ -345,6 +359,23 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_GenericAttribut
     static const int FIELD___error = 2;
 
     //--//
+
+};
+
+struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_NativeDevices_OnChip
+{
+    NANOCLR_NATIVE_DECLARE(NativeInitService___BOOLEAN);
+    NANOCLR_NATIVE_DECLARE(NativeStartAdvertising___BOOLEAN);
+    NANOCLR_NATIVE_DECLARE(NativeStopAdvertising___VOID);
+    NANOCLR_NATIVE_DECLARE(NativeNotifyClient___I4__U2__U2__SZARRAY_U1);
+    NANOCLR_NATIVE_DECLARE(NativeReadRespondWithValue___VOID__U2__SZARRAY_U1);
+    NANOCLR_NATIVE_DECLARE(NativeReadRespondWithProtocolError___VOID__U2__U1);
+    NANOCLR_NATIVE_DECLARE(NativeWriteGetData___SZARRAY_U1__U2);
+    NANOCLR_NATIVE_DECLARE(NativeWriteRespond___VOID__U2);
+    NANOCLR_NATIVE_DECLARE(NativeWriteRespondWithProtocolError___VOID__U2__U1);
+
+    //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_Spp_SppReceivedDataEventArgs
@@ -352,6 +383,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_Spp_SppReceived
     static const int FIELD___data = 1;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_Spp_NordicSpp
@@ -366,6 +398,7 @@ struct Library_sys_dev_ble_native_nanoFramework_Device_Bluetooth_Spp_NordicSpp
     static const int FIELD__ConnectedEvent = 8;
 
     //--//
+
 };
 
 struct Library_sys_dev_ble_native_ThisAssembly
@@ -373,8 +406,9 @@ struct Library_sys_dev_ble_native_ThisAssembly
     static const int FIELD_STATIC__GitCommitDate = 3;
 
     //--//
+
 };
 
 extern const CLR_RT_NativeAssemblyData g_CLR_AssemblyNative_nanoFramework_Device_Bluetooth;
 
-#endif //_SYS_DEV_BLE_NATIVE_H_
+#endif  //_SYS_DEV_BLE_NATIVE_H_
